# src/main/java/com/copyrightregister/db/copyrightregister_db/dtos/UserDto.java
package com.copyrightregister.db.copyrightregister_db.dtos;

import com.copyrightregister.db.copyrightregister_db.dtos.base.UserBaseDto;

public class UserDto extends UserBaseDto {

	//OVERRIDE HERE YOUR CUSTOMIZATIONS
}